UnitfulLinearAlgebra

Documentation for UnitfulLinearAlgebra.

Base.Matrix

LinearAlgebra.Diagonal

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.MultipliableMatrix

UnitfulLinearAlgebra.MultipliableMatrix

UnitfulLinearAlgebra.SquarableMatrix

UnitfulLinearAlgebra.UniformMatrix

UnitfulLinearAlgebra.UniformMatrix

UnitfulLinearAlgebra.UnitSymmetricMatrix

Base.:*

Base.:*

Base.:*

Base.:+

Base.:-

Base.:\

Base.getindex

Base.getproperty

Base.inv

Base.setindex!

Base.transpose

LinearAlgebra.cholesky

LinearAlgebra.det

LinearAlgebra.diagm

LinearAlgebra.lu

UnitfulLinearAlgebra.BestMultipliableMatrix

UnitfulLinearAlgebra.BestMultipliableMatrix

UnitfulLinearAlgebra.convert_unitdomain

UnitfulLinearAlgebra.convert_unitrange

UnitfulLinearAlgebra.dimensionless

UnitfulLinearAlgebra.domainlength

UnitfulLinearAlgebra.dottable

UnitfulLinearAlgebra.endomorphic

UnitfulLinearAlgebra.exact

UnitfulLinearAlgebra.identitymatrix

UnitfulLinearAlgebra.invdimension

UnitfulLinearAlgebra.left_uniform

UnitfulLinearAlgebra.multipliable

UnitfulLinearAlgebra.parallel

UnitfulLinearAlgebra.rangelength

UnitfulLinearAlgebra.right_uniform

UnitfulLinearAlgebra.similarity

UnitfulLinearAlgebra.uniform

Base.Matrix — Method

function Matrix(A::MultipliableMatrix)

Expand A into array form
Useful for tests, display
pp. 193, Hart

LinearAlgebra.Diagonal — Method

function Diagonal(v::AbstractVector,r::Unitful.Unitlike,d::Unitful.Unitlike; exact = false)

Construct diagonal matrix with units where the diagonal has elements `v`.
If `v` has units, check that they conform with dimensional unit range `r`
 and dimensional unit domain `d`.
Like `LinearAlgebra.Diagonal`, this extension is restricted to square matrices.

UnitfulLinearAlgebra.EndomorphicMatrix — Type

struct EndomorphicMatrix

An endomorphic matrix maps a dimensioned vector space
to itself. The dimensional range and domain are the same.

Attributes

numbers: numerical (dimensionless) matrix

unitrange: dimensional range in terms of units, and also equal the dimensional domain

exact: geometric (true) or algebraic (false) interpretation

UnitfulLinearAlgebra.EndomorphicMatrix — Method

function EndomorphicMatrix

Constructor with keyword argument `exact=false`.
If `exact` not specified, defaults to `false`.

UnitfulLinearAlgebra.EndomorphicMatrix — Method

 EndomorphicMatrix(A)

Transform array to EndomorphicMatrix type

UnitfulLinearAlgebra.EndomorphicMatrix — Method

function EndomorphicMatrix(A::T) where T <: Number

Special case of a scalar. Must be dimensionless.

UnitfulLinearAlgebra.MultipliableMatrix — Type

struct MultipliableMatrix

Matrices with units that a physically reasonable,
i.e., more than just an array of values with units.

Multipliable matrices have dimensions that are consistent with many linear algebraic manipulations, including multiplication.

Hart suggests that these matrices simply be called "matrices", and that matrices with dimensional values that cannot be multiplied should be called "arrays."

Attributes

numbers: numerical (dimensionless) matrix

unitrange: dimensional range in terms of units

unitdomain: dimensional domain in terms of units

exact: geometric (true) or algebraic (false) interpretation

UnitfulLinearAlgebra.MultipliableMatrix — Method

MultipliableMatrix(numbers,unitrange,unitdomain;exact=false)

Constructor where `exact` is a keyword argument. One may construct a MultipliableMatrix without specifying exact, in which case it defaults to `false`.

UnitfulLinearAlgebra.SquarableMatrix — Type

struct SquarableMatrix

An squarable matrix is one where 𝐀² is defined.
Definition: dimensional range and domain are parallel.

Attributes

numbers: numerical (dimensionless) matrix

unitrange: dimensional range in terms of units, this is also the domain

Δunitdomain: shift to range that gives the domain

exact: geometric (true) or algebraic (false) interpretation

UnitfulLinearAlgebra.UniformMatrix — Type

struct UniformMatrix

Uniform matrix

Attributes

numbers: numerical (dimensionless) matrix

unitrange: uniform dimensional range expressed as a single unit

unitdomain: uniform dimensional domain expressed as a single unit

exact: geometric (true) or algebraic (false) interpretation

UnitfulLinearAlgebra.UniformMatrix — Method

function UniformMatrix

Constructor with keyword argument `exact=false`.
If `exact` not specified, defaults to `false`.

UnitfulLinearAlgebra.UnitSymmetricMatrix — Type

struct UnitSymmetricMatrix

An squarable matrix is one where 𝐀² is defined.
Definition: inverse dimensional range and dimensional domain are parallel.
Called "dimensionally symmetric" by Hart.

Attributes

numbers: numerical (dimensionless) matrix

unitrange: dimensional range in terms of units, this is also the domain

Δunitdomain: shift to range that gives the domain

exact: geometric (true) or algebraic (false) interpretation

Base.:* — Method

function *(A::MultipliableMatrix,b)

Matrix-scalar multiplication with units/dimensions.
Must account for change in the unitrange when the
 scalar has units.
Here, take product of dimension of the scalar and the unitrange.
Alternatively, divide the domain by the dimension of the scalar. 
Matrix-scalar multiplication is commutative.
Result is `exact` if input matrix is exact and scalar is dimensionless. 
Note: special matrix forms revert to a product that is a MultipliableMatrix.

Base.:* — Method

function *(A,B)

Matrix-matrix multiplication with units/dimensions.
A*B represents two successive transformations.
Unitrange of B should equal domain of A in geometric interpretation.
Unitrange of B should be parallel to unitdomain of A in algebraic interpretation.

Note: special matrix forms revert to a product that is a MultipliableMatrix.

Base.:* — Method

function *(A::AbstractMultipliableMatrix,b)

Matrix-vector multiplication with units/dimensions.
Unitful also handles this case, but here there is added
efficiency in the storage of units/dimensions by accounting
for the necessary structure of the matrix.

Base.:+ — Method

function +(A,B)

Matrix-matrix addition with units/dimensions.
A+B requires the two matrices to have dimensional similarity.

Base.:- — Method

function -(A,B)

Matrix-matrix subtraction with units/dimensions.
A-B requires the two matrices to have dimensional similarity.

Base.:\ — Method

 function ldiv!

 Left divide of Multipliable Matrix.
 Reverse mapping from unitdomain to range.
 Is `exact` if input is exact.

Base.getindex — Method

function getindex(A::MultipliableMatrix,i::Integer,j::Integer)

Recover element (i,j) of a MultipliableMatrix.
Part of the AbstractArray interface.

#Input

A::MultipliableMatrix

i::Integer: row index

j::Integer: column index

#Output

Quantity: numerical value and units

Base.getproperty — Method

function getproperty(F::LU{T,<:AbstractMultipliableMatrix,Vector{Int64}}, d::Symbol) where T

Extend LinearAlgebra.getproperty for AbstractMultipliableMatrix.

LU factorization stores L and U together.
Extract L and U while keeping consistent
with dimensional domain and range.

Base.inv — Method

 function inv

 Inverse of Multipliable Matrix.
 Only defined for nonsingular matrices.
 Inverse reverses mapping from unitdomain to range.
 Is `exact` if input is exact.

Hart, pp. 205.

Base.setindex! — Method

function setindex!(A::MultipliableMatrix,v,i,j)

Set element (i,j) of a MultipliableMatrix.
Part of the AbstractArray interface.

#Input

A::MultipliableMatrix

v: new value

i::Integer: row index

j::Integer: column index

#Output

Quantity: numerical value and units

Base.transpose — Method

function transpose

Defined by condition `A[i,j] = transpose(A)[j,i]`.
Not analogous to function for dimensionless matrices.

Hart, pp. 205.

LinearAlgebra.cholesky — Method

function cholesky(A::AbstractMultipliableMatrix)

Cholesky decomposition extended for matrices with units.
Requires unit (or dimensionally) symmetric matrix.

LinearAlgebra.det — Method

function det

LinearAlgebra.diagm — Method

function diagm(v::AbstractVector,r::Unitful.Unitlike,d::Unitful.Unitlike; exact = false)

Construct diagonal matrix with units where the diagonal has elements `v`.
If `v` has units, check that they conform with dimensional unit range `r`
 and dimensional unit domain `d`. Works for square or non-square matrices.

LinearAlgebra.lu — Method

function lu(A::AbstractMultipliableMatrix{T})

Extend `lu` factorization to AbstractMultipliableMatrix.
Related to Gaussian elimination.
Store dimensional domain and range in "factors" attribute
even though this is not truly a MultipliableMatrix.
Returns `LU` type in analogy with `lu` for unitless matrices.
Based on LDU factorization, Hart, pp. 204.

UnitfulLinearAlgebra.BestMultipliableMatrix — Method

function BestMultipliableMatrix(numbers,unitrange,unitdomain;exact=false)

What kind of Multipliable Matrix is the best representation?

UnitfulLinearAlgebra.BestMultipliableMatrix — Method

 BestMultipliableMatrix(A::Matrix)

Transform array to a type <: AbstractMultipliableMatrix.
Finds best representation amongst
UniformMatrix, EndomorphicMatrix, or MultipliableMatrix.
Assumes `exact=false`

UnitfulLinearAlgebra.convert_unitdomain — Method

function convert_unitdomain(A, newdomain)

When using the geometric interpretation of matrices,
it is useful to convert the dimensional domain of the
matrix to match the expected vectors during multiplication.
Here we set the matrix to `exact=true` after this step.

UnitfulLinearAlgebra.convert_unitrange — Method

function convert_unitrange(A, newrange)

When using the geometric interpretation of matrices,
it is useful to convert the dimensional range of the
matrix to match the desired output of multiplication.
Here we set the matrix to `exact=true` after this step.

UnitfulLinearAlgebra.dimensionless — Method

 function dimensionless(A)

 Not all dimensionless matrices have
 dimensionless domain and range.

UnitfulLinearAlgebra.domainlength — Method

function domainlength(A::MultipliableMatrix)

Numerical dimension (length or size) of unitdomain of A

UnitfulLinearAlgebra.dottable — Method

function dottable(a,b)

Are two quantities dimensionally compatible
to take a dot product?

UnitfulLinearAlgebra.endomorphic — Method

function endomorphic(A)::Bool

Endomorphic matrices have a particular structure
 of the units/dimensions in the array.

UnitfulLinearAlgebra.exact — Method

function exact(A)

exact=true: geometric interpretation of unitdomain and unitrange

exact=false: algebraic interpretation

UnitfulLinearAlgebra.identitymatrix — Method

function identitymatrix(dimrange)

Input: dimensional (unit) range.
`A + I` only defined when `endomorphic(A)=true`
When accounting for units, there are many identity matrices.
This function returns a particular identity matrix
defined by its dimensional range.
Hart, pp. 200.

UnitfulLinearAlgebra.invdimension — Method

function invdimension

Dimensional inverse
  
pp. 64, Hart, `a~` in his notation

UnitfulLinearAlgebra.left_uniform — Method

function left_uniform(A)

Does the unitrange of A have uniform dimensions?

UnitfulLinearAlgebra.multipliable — Method

function multipliable(A)::Bool

Is an array multipliable?
It requires a particular structure of the units/dimensions in the array.

UnitfulLinearAlgebra.parallel — Method

function parallel

Vector a is dimensionally parallel to vector b if
they have the same length and a consistent dimensional
change relates corresponding components.
Guaranteed if two vectors are dimensionally similar.
True for scalars in all cases. 

pp. 188, Hart
Note: Hart uses ≈, but this conflicts with an existing Julia function.

UnitfulLinearAlgebra.rangelength — Method

function rangelength(A::MultipliableMatrix)

Numerical dimension (length or size) of unitrange

UnitfulLinearAlgebra.right_uniform — Method

function right_uniform(A)

Does the unitdomain of A have uniform dimensions?

UnitfulLinearAlgebra.similarity — Method

function similarity(a,b)::Bool

Dimensional similarity of vectors, a binary relation
Read "a has the same dimensional form as b"
`a` and `b` may still have different units.
A stronger condition than being parallel.
pp. 184, Hart

UnitfulLinearAlgebra.uniform — Method

function uniform(a)

Is the dimension of this quantity uniform?

There must be a way to inspect the Unitful type to answer this.