UnitfulLinearAlgebra

Documentation for UnitfulLinearAlgebra.

Base.Matrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.EndomorphicMatrix

UnitfulLinearAlgebra.MultipliableMatrix

UnitfulLinearAlgebra.MultipliableMatrix

UnitfulLinearAlgebra.MultipliableMatrix

UnitfulLinearAlgebra.SquarableMatrix

Base.:*

Base.:*

Base.:*

Base.inv

Base.similar

LinearAlgebra.det

LinearAlgebra.lu

UnitfulLinearAlgebra.convert_domain

UnitfulLinearAlgebra.convert_range

UnitfulLinearAlgebra.dimensionless

UnitfulLinearAlgebra.domainlength

UnitfulLinearAlgebra.dottable

UnitfulLinearAlgebra.element

UnitfulLinearAlgebra.endomorphic

UnitfulLinearAlgebra.exact

UnitfulLinearAlgebra.invdimension

UnitfulLinearAlgebra.left_uniform

UnitfulLinearAlgebra.multipliable

UnitfulLinearAlgebra.parallel

UnitfulLinearAlgebra.rangelength

UnitfulLinearAlgebra.right_uniform

UnitfulLinearAlgebra.uniform

Base.Matrix — Method

function Matrix(A::MultipliableMatrix)

Expand A into array form
Useful for tests, display
pp. 193, Hart

UnitfulLinearAlgebra.EndomorphicMatrix — Type

struct EndomorphicMatrix

An endomorphic matrix maps a dimensioned vector space
to itself. The dimensional range and domain are the same.

Attributes

numbers: numerical (dimensionless) matrix

range: dimensional range in terms of units, this is also the domain

UnitfulLinearAlgebra.EndomorphicMatrix — Method

function EndomorphicMatrix

Constructor where `exact` is a keyword argument. One may construct an EndomorphicMatrix without specifying exact, in which case it defaults to `false`.

UnitfulLinearAlgebra.EndomorphicMatrix — Method

 EndomorphicMatrix(A)

Transform array to EndomorphicMatrix type

UnitfulLinearAlgebra.EndomorphicMatrix — Method

function EndomorphicMatrix(A::T) where T <: Number

Special case of a scalar. Must be dimensionless.

UnitfulLinearAlgebra.MultipliableMatrix — Type

struct MultipliableMatrix

Matrices with units that a physically reasonable,
i.e., more than just an array of values with units.

Multipliable matrices have dimensions that are consistent with many linear algebraic manipulations, including multiplication.

Hart suggests that these matrices simply be called "matrices", and that matrices with dimensional values that cannot be multiplied should be called "arrays."

Attributes

numbers: numerical (dimensionless) matrix

range: dimensional range in terms of units

domain: dimensional domain in terms of units

exact: geometric (true) or algebraic (false) interpretation

UnitfulLinearAlgebra.MultipliableMatrix — Method

MultipliableMatrix(numbers,range,domain;exact=false)

Constructor where `exact` is a keyword argument. One may construct a MultipliableMatrix without specifying exact, in which case it defaults to `false`.

UnitfulLinearAlgebra.MultipliableMatrix — Method

 MultipliableMatrix(array)

Transform array to MultipliableMatrix

UnitfulLinearAlgebra.SquarableMatrix — Type

struct SquarableMatrix

An squarable matrix is one where 𝐀² is defined.
It is the case if the dimensional range and domain are parallel.

Attributes

numbers: numerical (dimensionless) matrix

range: dimensional range in terms of units, this is also the domain

domainshift: shift to range that gives the domain

Base.:* — Method

function *(A::MultipliableMatrix,b)

Matrix-scalar multiplication with units/dimensions.
Must account for change in the range when the
 scalar has units.
Here, take product of dimension of the scalar and the range.
Alternatively, divide the domain by the dimension of the scalar. 
Matrix-scalar multiplication is commutative.
Result is `exact` if input matrix is exact and scalar is dimensionless. 
Note: special matrix forms revert to a product that is a MultipliableMatrix.

Base.:* — Method

function *(A,B)

Matrix-matrix multiplication with units/dimensions.
A*B represents two successive transformations.
Range of B should equal domain of A in geometric interpretation.
Range of B should be parallel to domain of A in algebraic interpretation.

Note: special matrix forms revert to a product that is a MultipliableMatrix.

Base.:* — Method

function *(A::MultipliableMatrix,b)

Matrix-vector multiplication with units/dimensions.
Unitful also handles this case, but here there is added
efficiency in the storage of units/dimensions by accounting
for the necessary structure of the matrix.

Base.inv — Method

 function inv

Base.similar — Method

function similar(a,b)::Bool

Dimensional similarity of vectors, a binary relation
Read "a has the same dimensional form as b"
`a` and `b` may still have different units.
A stronger condition than being parallel.
pp. 184, Hart

LinearAlgebra.det — Method

function det

LinearAlgebra.lu — Method

function lu(A::MultipliableMatrix{Float64})

UnitfulLinearAlgebra.convert_domain — Method

function convert_domain(A, newdomain)

When using the geometric interpretation of matrices,
it is useful to convert the dimensional domain of the
matrix to match the expected vectors during multiplication.
Here we set the matrix to `exact=true` after this step.

UnitfulLinearAlgebra.convert_range — Method

function convert_range(A, newrange)

When using the geometric interpretation of matrices,
it is useful to convert the dimensional range of the
matrix to match the desired output of multiplication.
Here we set the matrix to `exact=true` after this step.

UnitfulLinearAlgebra.dimensionless — Method

 function dimensionless(A)

 Not all dimensionless matrices have
 dimensionless domain and range.

UnitfulLinearAlgebra.domainlength — Method

function domainlength(A::MultipliableMatrix)

Numerical dimension (length or size) of domain of A

UnitfulLinearAlgebra.dottable — Method

function dottable(a,b)

Are two quantities dimensionally compatible
to take a dot product?

UnitfulLinearAlgebra.element — Method

function element(A::MultipliableMatrix,i::Integer,j::Integer)

Recover element (i,j) of a MultipliableMatrix.

#Input

A::MultipliableMatrix

i::Integer: row index

j::Integer: column index

#Output

Quantity: numerical value and units

UnitfulLinearAlgebra.endomorphic — Method

function endomorphic(A)::Bool

Is an array endomorphic?
It requires a particular structure of the units/dimensions in the array.

UnitfulLinearAlgebra.exact — Method

function exact(A)

exact=true: geometric interpretation of domain and range

exact=false: algebraic interpretation

UnitfulLinearAlgebra.invdimension — Method

function invdimension

Dimensional inverse
  
pp. 64, Hart, `a~` in his notation

UnitfulLinearAlgebra.left_uniform — Method

function left_uniform(A)

Does the range of A have uniform dimensions?

UnitfulLinearAlgebra.multipliable — Method

function multipliable(A)::Bool

Is an array multipliable?
It requires a particular structure of the units/dimensions in the array.

UnitfulLinearAlgebra.parallel — Method

function parallel

Vector a is dimensionally parallel to vector b if
they have the same length and a consistent dimensional
change relates corresponding components.
Guaranteed if two vectors are dimensionally similar.
True for scalars in all cases. 

pp. 188, Hart
Note: Hart uses ≈, but this conflicts with an existing Julia function.

UnitfulLinearAlgebra.rangelength — Method

function rangelength(A::MultipliableMatrix)

Numerical dimension (length or size) of range

UnitfulLinearAlgebra.right_uniform — Method

function right_uniform(A)

Does the domain of A have uniform dimensions?

UnitfulLinearAlgebra.uniform — Method

function uniform(a)

Is the dimension of this quantity uniform?

There must be a way to inspect the Unitful type to answer this.